feat: add 'TextLayerEventsController'
