+++ packages/services/src/controllers/index.ts
@@ -1,2 +1,3 @@
 export * from './KeyboardShortcutController/KeyboardShortcutController';
 export * from './SelectionEventsController/SelectionEventsController';
+export * from './TextLayerEventsController/TextLayerEventsController';
